:rocket: Do not use customFields for now

+++ src/zeus/const.ts
@@ -1772,7 +1772,7 @@ export const ReturnTypes: Record<string,any> = {
 		NIP:"String"
 	},
 	ProductVariantCustomFields:{
-		beforePrice:"Float"
+		beforePrice:"Int"
 	}
 }
